Fix Countable[Map[K, V]], move towards Countable[A => B].
The previous strategy for enumerating all map values was
fundamentally busted. Our new strategy is easier to show
correct and doesn't seem to have gaps or duplicates.

In the same vein as Countable[Map[K, V]], I've started
laying the ground-work for Countable function instances.
Using kronecker.Functional we could write a method
generating Countable[Z => B] given a Finite[B].

To create Countable[A => B] for other types we need
some kind of Indexable[A] instance that allows us to
go from an A value to its index. Then we could implement
(Indexable[A], Finite[B]) => Countable[A => B].

For both maps and functions, the best solution for dealing
with an infinite value type continues to elude me.